requirements.txt
fastapi
uvicorn
pydantic
qdrant-client
sentence-transformers
openai
python-dotenv
